+++ lib/postman.js
@@ -1,40 +1,13 @@
 const env = require('./env')
 const scheduler = require('./scheduler')
+const {
+  locationSelect,
+  cancelButton,
+  cancelDateButton,
+  updateButton
+} = require('./components')
 
 const random = (list) => list[Math.floor(Math.random() * list.length)]
-
-const locationSelect = {
-  name: 'location',
-  type: 'select',
-  text: 'Location',
-  options: env.locations.list.map((location) => ({
-    text: location,
-    value: location
-  }))
-}
-
-const cancelButton = {
-  name: 'cancel',
-  type: 'button',
-  text: 'Cancel',
-  style: 'danger',
-  value: true
-}
-
-const cancelDateButton = {
-  name: 'cancelDate',
-  type: 'button',
-  text: 'Cancel Date',
-  style: 'danger',
-  value: true
-}
-
-const updateButton = {
-  name: 'update',
-  type: 'button',
-  text: 'Update',
-  value: true
-}
 
 class Postman {
   constructor () {
